Refactor KeyframeTrackClip into multiple files
