+++ src/timeline/keyframe-track-clip.c
@@ -1,21 +1,5 @@
 #include "keyframe-track-clip.h"
 
-struct _KeyframeTrackClip
-{
-    KeyframeTrack parent_instance;
-
-    GtkWidget *clip;
-    gboolean drag_active;
-    float drag_start_x;
-    float drag_current_x;
-
-    // TODO: Make properties?
-    float clip_timestamp_start;
-    float clip_timestamp_end;
-
-    float start_position;
-};
-
 G_DEFINE_FINAL_TYPE (KeyframeTrackClip, keyframe_track_clip, KEYFRAME_TYPE_TRACK)
 
 enum {
@@ -26,141 +10,6 @@ enum {
 };
 
 static GParamSpec *properties [N_PROPS];
-
-
-
-// START CLIP WIDGET
-
-
-#define KEYFRAME_TYPE_TIMELINE_CLIP (keyframe_timeline_clip_get_type())
-
-G_DECLARE_FINAL_TYPE (KeyframeTimelineClip, keyframe_timeline_clip, KEYFRAME, TIMELINE_CLIP, GtkWidget)
-
-struct _KeyframeTimelineClip
-{
-    GtkWidget parent_instance;
-};
-
-G_DEFINE_FINAL_TYPE (KeyframeTimelineClip, keyframe_timeline_clip, GTK_TYPE_WIDGET)
-
-static GtkWidget *
-keyframe_timeline_clip_new ()
-{
-    return GTK_WIDGET (g_object_new (KEYFRAME_TYPE_TIMELINE_CLIP, 0));
-}
-
-static void
-keyframe_timeline_clip_class_init (KeyframeTimelineClipClass *klass)
-{
-    GtkWidgetClass *widget_class = GTK_WIDGET_CLASS (klass);
-
-    gtk_widget_class_set_css_name (widget_class, "clip");
-    gtk_widget_class_set_layout_manager_type (widget_class, GTK_TYPE_BOX_LAYOUT);
-}
-
-static void
-keyframe_timeline_clip_set_active_cursor (KeyframeTimelineClip *self, gboolean pressed)
-{
-    gchar *cursor_name = pressed ? "grabbing" : "grab";
-    GdkCursor *cursor = gdk_cursor_new_from_name (cursor_name, NULL);
-    gtk_widget_set_cursor (GTK_WIDGET (self), cursor);
-}
-
-static void
-keyframe_timeline_clip_init (KeyframeTimelineClip *self)
-{
-    // TODO: Drag Handles
-    keyframe_timeline_clip_set_active_cursor (self, FALSE);
-}
-
-
-// END CLIP WIDGET
-
-// START LAYOUT MANAGER
-
-
-#define KEYFRAME_TYPE_TRACK_CLIP_LAYOUT (keyframe_track_clip_layout_get_type())
-
-G_DECLARE_FINAL_TYPE (KeyframeTrackClipLayout, keyframe_track_clip_layout, KEYFRAME, TRACK_CLIP_LAYOUT, GtkLayoutManager)
-
-struct _KeyframeTrackClipLayout
-{
-    GtkLayoutManager parent_instance;
-};
-
-G_DEFINE_FINAL_TYPE (KeyframeTrackClipLayout, keyframe_track_clip_layout, GTK_TYPE_LAYOUT_MANAGER)
-
-static void
-keyframe_track_clip_layout_measure (GtkLayoutManager *layout_manager,
-                                             GtkWidget        *widget,
-                                             GtkOrientation    orientation,
-                                             int               for_size,
-                                             int              *minimum,
-                                             int              *natural,
-                                             int              *minimum_baseline,
-                                             int              *natural_baseline)
-{
-    *minimum = 0;
-    *natural = 0;
-}
-
-static void
-keyframe_track_clip_layout_allocate (GtkLayoutManager *layout_manager,
-                                              GtkWidget        *widget,
-                                              int               width,
-                                              int               height,
-                                              int               baseline)
-{
-    KeyframeTrackClip *clip_track = KEYFRAME_TRACK_CLIP (widget);
-
-    if (!gtk_widget_should_layout (clip_track->clip))
-        return;
-
-    float clip_width = clip_track->clip_timestamp_end - clip_track->clip_timestamp_start;
-
-    if (clip_track->drag_active)
-    {
-        gtk_widget_size_allocate (clip_track->clip,
-                              &(const GtkAllocation){ clip_track->drag_current_x - clip_track->start_position, 0, clip_width, height },
-                              -1);
-    }
-    else
-    {
-        g_print ("%f %f\n", clip_track->clip_timestamp_start, clip_track->clip_timestamp_end);
-        gtk_widget_size_allocate (clip_track->clip,
-                              &(const GtkAllocation){ clip_track->clip_timestamp_start - clip_track->start_position, 0, clip_width, height },
-                              -1);
-    }
-}
-
-static GtkSizeRequestMode
-keyframe_track_clip_layout_get_request_mode (GtkLayoutManager *layout_manager,
-                                                      GtkWidget        *widget)
-{
-    return GTK_SIZE_REQUEST_CONSTANT_SIZE;
-}
-
-static void
-keyframe_track_clip_layout_class_init (KeyframeTrackClipLayoutClass *klass)
-{
-    GtkLayoutManagerClass *layout_manager_class = GTK_LAYOUT_MANAGER_CLASS (klass);
-
-    layout_manager_class->measure = keyframe_track_clip_layout_measure;
-    layout_manager_class->allocate = keyframe_track_clip_layout_allocate;
-    layout_manager_class->get_request_mode = keyframe_track_clip_layout_get_request_mode;
-}
-
-static void
-keyframe_track_clip_layout_init (KeyframeTrackClipLayout *self)
-{
-
-}
-
-
-// END LAYOUT MANAGER
-
-
-
 
 GtkWidget *
 keyframe_track_clip_new (void)
@@ -294,7 +143,7 @@ cb_drag_begin (GtkGestureDrag            *gesture,
     self->drag_current_x = self->clip_timestamp_start;
     self->drag_active = true;
 
-    keyframe_timeline_clip_set_active_cursor (KEYFRAME_TIMELINE_CLIP (self->clip), TRUE);
+    keyframe_track_clip_widget_set_active_cursor (KEYFRAME_TRACK_CLIP_WIDGET (self->clip), TRUE);
 }
 
 static void
@@ -329,7 +178,7 @@ cb_drag_end (GtkGestureDrag            *gesture,
     self->drag_start_x = 0;
     self->drag_current_x = 0;
 
-    keyframe_timeline_clip_set_active_cursor (KEYFRAME_TIMELINE_CLIP (self->clip), FALSE);
+    keyframe_track_clip_widget_set_active_cursor (KEYFRAME_TRACK_CLIP_WIDGET (self->clip), FALSE);
 
     gtk_widget_queue_allocate (GTK_WIDGET (self));
 }
@@ -337,7 +186,7 @@ cb_drag_end (GtkGestureDrag            *gesture,
 static void
 keyframe_track_clip_init (KeyframeTrackClip *self)
 {
-    self->clip = keyframe_timeline_clip_new ();
+    self->clip = keyframe_track_clip_widget_new ();
     gtk_widget_set_parent (self->clip, GTK_WIDGET (self));
 
     // TODO: Scaling/Zoom
